+++ xmlschema/validators/elements.py
@@ -14,19 +14,21 @@ import warnings
 from decimal import Decimal
 from typing import Optional
 from elementpath import XPath2Parser, ElementPathError, XPathContext
-from elementpath.datatypes import AbstractDateTime, Duration
+from elementpath.datatypes import AbstractDateTime, Duration, AbstractBinary
 
 from ..exceptions import XMLSchemaTypeError, XMLSchemaValueError
 from ..names import XSD_COMPLEX_TYPE, XSD_SIMPLE_TYPE, XSD_ALTERNATIVE, \
     XSD_ELEMENT, XSD_ANY_TYPE, XSD_UNIQUE, XSD_KEY, XSD_KEYREF, XSI_NIL, \
     XSI_TYPE, XSD_ERROR, XSD_NOTATION_TYPE
 from ..etree import etree_element
-from ..helpers import get_qname, get_namespace, etree_iter_location_hints
+from ..helpers import get_qname, get_namespace, etree_iter_location_hints, \
+    raw_xml_encode, strictly_equal
+from .. import dataobjects
 from ..converters import ElementData, XMLSchemaConverter
 from ..xpath import XMLSchemaProxy, ElementPathMixin
 
 from .exceptions import XMLSchemaValidationError, XMLSchemaTypeTableWarning
-from .helpers import get_xsd_derivation_attribute, raw_xml_encode, strictly_equal
+from .helpers import get_xsd_derivation_attribute
 from .xsdbase import XSD_TYPE_DERIVATIONS, XSD_ELEMENT_DERIVATIONS, \
     XsdComponent, XsdType, ValidationMixin
 from .particles import ParticleMixin
@@ -76,6 +78,8 @@ class XsdElement(XsdComponent, ValidationMixin, ParticleMixin, ElementPathMixin)
     _block = None
     _final = None
     _head_type = None
+
+    binding = None
 
     def __init__(self, elem, schema, parent):
         super(XsdElement, self).__init__(elem, schema, parent)
@@ -375,6 +379,23 @@ class XsdElement(XsdComponent, ValidationMixin, ParticleMixin, ElementPathMixin)
             return self._block
         return self.schema.block_default
 
+    def get_binding(self, *bases, replace_existing=False, **attrs):
+        """
+        Gets data object binding for XSD element, creating a new one if it doesn't exist.
+
+        :param bases: base classes to use for creating the binding class.
+        :param replace_existing: provide `True` to replace an existing binding class.
+        :param attrs: attribute and method definitions for the binding class body.
+        """
+        if self.binding is None or replace_existing:
+            if not bases:
+                bases = (dataobjects.DataElement,)
+            attrs['xsd_element'] = self
+            class_name = '{}Binding'.format(self.local_name.title().replace('_', ''))
+            self.binding = dataobjects.DataBindingMeta(class_name, bases, attrs)
+
+        return self.binding
+
     def get_attribute(self, name):
         if name[0] != '{':
             return self.type.attributes[get_qname(self.type.target_namespace, name)]
@@ -630,7 +651,7 @@ class XsdElement(XsdComponent, ValidationMixin, ParticleMixin, ElementPathMixin)
 
             if self.fixed is not None and \
                     (len(elem) > 0 or value is not None and self.fixed != value):
-                reason = "must have the fixed value %r." % self.fixed
+                reason = "must have the fixed value %r" % self.fixed
                 yield self.validation_error(validation, reason, elem, **kwargs)
 
         else:
@@ -646,7 +667,7 @@ class XsdElement(XsdComponent, ValidationMixin, ParticleMixin, ElementPathMixin)
                     pass
                 elif not strictly_equal(xsd_type.text_decode(text),
                                         xsd_type.text_decode(self.fixed)):
-                    reason = "must have the fixed value %r." % self.fixed
+                    reason = "must have the fixed value %r" % self.fixed
                     yield self.validation_error(validation, reason, elem, **kwargs)
 
             elif not text and self.default is not None and kwargs.get('use_defaults'):
@@ -686,20 +707,22 @@ class XsdElement(XsdComponent, ValidationMixin, ParticleMixin, ElementPathMixin)
                     else:
                         value = result
 
-            if isinstance(value, Decimal):
+            if isinstance(value, (int, float, list)) or value is None:
+                pass
+            elif isinstance(value, str):
+                if value.startswith('{') and xsd_type.is_qname():
+                    value = text
+            elif isinstance(value, Decimal):
                 try:
                     value = kwargs['decimal_type'](value)
                 except (KeyError, TypeError):
                     pass
             elif isinstance(value, (AbstractDateTime, Duration)):
-                try:
-                    if kwargs['datetime_types'] is not True:
-                        value = elem.text
-                except KeyError:
+                if not kwargs.get('datetime_types'):
                     value = elem.text
-            elif isinstance(value, str) and value.startswith('{'):
-                if xsd_type.is_qname():
-                    value = text
+            elif isinstance(value, AbstractBinary):
+                if not kwargs.get('binary_types'):
+                    value = elem.text
 
         if converter is not None:
             element_data = ElementData(elem.tag, value, content, attributes)
@@ -765,6 +788,8 @@ class XsdElement(XsdComponent, ValidationMixin, ParticleMixin, ElementPathMixin)
         :return: yields an Element, eventually preceded by a sequence of \
         validation or encoding errors.
         """
+        errors = []
+
         try:
             converter = kwargs['converter']
         except KeyError:
@@ -777,10 +802,17 @@ class XsdElement(XsdComponent, ValidationMixin, ParticleMixin, ElementPathMixin)
             level = kwargs['level']
         except KeyError:
             level = 0
+            element_data = converter.element_encode(obj, self, level)
+            if not self.is_matching(element_data.tag, self.default_namespace):
+                errors.append("data tag does not match XSD element name")
 
-        element_data = converter.element_encode(obj, self, level)
-        errors = []
-        tag = element_data.tag
+            if 'max_depth' in kwargs and kwargs['max_depth'] == 0:
+                for e in errors:
+                    yield self.validation_error(validation, e, **kwargs)
+                return
+        else:
+            element_data = converter.element_encode(obj, self, level)
+
         text = None
         children = element_data.content
         attributes = ()
@@ -872,7 +904,7 @@ class XsdElement(XsdComponent, ValidationMixin, ParticleMixin, ElementPathMixin)
                 elif result:
                     text, children = result
 
-        elem = converter.etree_element(tag, text, children, attributes, level)
+        elem = converter.etree_element(element_data.tag, text, children, attributes, level)
 
         if errors:
             for e in errors: