+++ xmlschema/xpath.py
@@ -140,6 +140,7 @@ class XMLSchemaProxy(AbstractSchemaProxy):
         return self._schema.find(path, namespaces)
 
     def is_instance(self, obj, type_qname):
+        # FIXME: use elementpath.datatypes for checking atomic datatypes
         xsd_type = self._schema.maps.types[type_qname]
         try:
             xsd_type.encode(obj)
@@ -261,7 +262,7 @@ class ElementPathMixin(Sequence):
 
         :param path: an XPath expression that considers the XSD component as the root element.
         :param namespaces: an optional mapping from namespace prefix to namespace URI.
-        :return: The first matching XSD subelement or ``None`` if there is not match.
+        :return: the first matching XSD subelement or ``None`` if there is no match.
         """
         context = XMLSchemaContext(self)
         return next(self._xpath_parse(path, namespaces).select_results(context), None)